wip: embeddings and image generations handlers
