+++ atoma-service/Cargo.toml
@@ -39,6 +39,7 @@ tracing-appender = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter", "json", "local-time"] }
 utoipa = { workspace = true, features = ["axum_extras"] }
 utoipa-swagger-ui = { workspace = true, features = ["axum"] }
+serde_yaml = "0.9.34"
 
 [dev-dependencies]
 rand = { workspace = true }
