+++ atoma-service/src/handlers/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod chat_completions;
+pub(crate) mod embeddings;
+pub(crate) mod image_generations;
