+++ atoma-service/src/server.rs
@@ -2,51 +2,29 @@ use std::sync::Arc;
 
 use atoma_state::types::AtomaAtomaStateManagerEvent;
 use axum::{
-    extract::State,
-    http::StatusCode,
     middleware::{from_fn, from_fn_with_state},
     response::IntoResponse,
     routing::{get, post},
-    Extension, Json, Router,
+    Json, Router,
 };
 use blake2::{
     digest::generic_array::{typenum::U32, GenericArray},
     Digest,
 };
 use flume::Sender as FlumeSender;
-use reqwest::Client;
 use serde_json::{json, Value};
 use sui_keys::keystore::FileBasedKeystore;
 use tokenizers::Tokenizer;
 use tokio::{net::TcpListener, signal, sync::watch::Sender};
 use tower::ServiceBuilder;
-use tracing::{error, info, instrument};
+use tracing::info;
 use utoipa::OpenApi;
 
 use crate::{
-    middleware::{signature_verification_middleware, verify_stack_permissions, RequestMetadata},
-    types::{ChatCompletionsRequest, ChatCompletionsResponse},
+    components::openapi::openapi_routes,
+    handlers::chat_completions::{chat_completions_handler, CHAT_COMPLETIONS_PATH},
+    middleware::{signature_verification_middleware, verify_stack_permissions},
 };
-
-const CHAT_COMPLETIONS_PATH: &str = "/v1/chat/completions";
-const HEALTH_PATH: &str = "/health";
-
-/// OpenAPI documentation for the chat completions endpoint
-#[derive(OpenApi)]
-#[openapi(
-    paths(
-        health,
-        chat_completions_handler
-    ),
-    components(
-        schemas(ChatCompletionsRequest, ChatCompletionsResponse)
-    ),
-    tags(
-        (name = "health", description = "Health check endpoint"),
-        (name = "chat-completions", description = "Chat completions endpoint")
-    )
-)]
-pub struct OpenApiDoc;
 
 /// Represents the shared state of the application.
 ///
@@ -123,6 +101,7 @@ pub struct AppState {
 pub fn create_router(app_state: AppState) -> Router {
     Router::new()
         .route(CHAT_COMPLETIONS_PATH, post(chat_completions_handler))
+        .merge(openapi_routes())
         .layer(
             ServiceBuilder::new()
                 .layer(from_fn(signature_verification_middleware))
@@ -208,137 +187,22 @@ pub async fn run_server(
 /// ```rust,ignore
 /// app.route("/health", get(health_check))
 /// ```
+
+pub const HEALTH_PATH: &str = "/health";
+#[derive(OpenApi)]
+#[openapi(paths(health))]
+pub(crate) struct HealthOpenApi;
+
 #[utoipa::path(
     get,
-    path = "/health",
+    path = HEALTH_PATH,
     tag = "health",
     responses(
-        (status = 200, description = "Service is healthy", body = Value)
+        (status = OK, description = "Service is healthy", body = Value)
     )
 )]
 pub async fn health() -> impl IntoResponse {
     Json(json!({ "status": "ok" }))
-}
-
-/// Handles chat completion requests by forwarding them to the inference service and managing token usage.
-///
-/// This handler performs several key operations:
-/// 1. Forwards the chat completion request to the inference service
-/// 2. Signs the response using the node's keystore
-/// 3. Tracks token usage for the stack
-///
-/// # Arguments
-///
-/// * `Extension((stack_small_id, estimated_total_tokens))` - Stack ID and estimated token count from middleware
-/// * `state` - Application state containing the inference client and keystore
-/// * `payload` - The chat completion request body
-///
-/// # Returns
-///
-/// Returns a JSON response containing:
-/// - The inference service's response
-/// - A cryptographic signature of the response
-///
-/// # Errors
-///
-/// Returns a `StatusCode::INTERNAL_SERVER_ERROR` if:
-/// - The inference service request fails
-/// - Response parsing fails
-/// - Response signing fails
-/// - Token usage update fails
-#[utoipa::path(
-    post,
-    path = "/v1/chat/completions",
-    tag = "chat",
-    request_body = ChatCompletionsRequest,
-    responses(
-        (status = 200, description = "Chat completion successful", body = ChatCompletionsResponse),
-        (status = 500, description = "Internal server error")
-    )
-)]
-#[instrument(
-    level = "info",
-    skip(state, payload),
-    fields(path = CHAT_COMPLETIONS_PATH)
-)]
-pub async fn chat_completions_handler(
-    Extension(request_metadata): Extension<RequestMetadata>,
-    State(state): State<AppState>,
-    Json(payload): Json<Value>,
-) -> Result<Json<Value>, StatusCode> {
-    let RequestMetadata {
-        stack_small_id,
-        estimated_total_tokens,
-        payload_hash,
-    } = request_metadata;
-    let client = Client::new();
-    let response = client
-        .post(format!(
-            "{}{}",
-            state.inference_service_url, CHAT_COMPLETIONS_PATH
-        ))
-        .json(&payload)
-        .send()
-        .await
-        .map_err(|e| {
-            error!("Error sending request to inference service: {}", e);
-            StatusCode::INTERNAL_SERVER_ERROR
-        })?;
-    let mut response_body = response.json::<Value>().await.map_err(|e| {
-        error!("Error reading response body: {}", e);
-        StatusCode::INTERNAL_SERVER_ERROR
-    })?;
-
-    // Sign the response body byte content and add the base64 encoded signature to the response body
-    let (response_hash, signature) =
-        utils::sign_response_body(&response_body, &state.keystore, state.address_index).map_err(
-            |e| {
-                error!("Error signing response body: {}", e);
-                StatusCode::INTERNAL_SERVER_ERROR
-            },
-        )?;
-    response_body["signature"] = json!(signature);
-
-    // Extract the response total number of tokens
-    let total_tokens = response_body
-        .get("usage")
-        .and_then(|usage| usage.get("total_tokens"))
-        .and_then(|total_tokens| total_tokens.as_u64())
-        .map(|n| n as i64)
-        .unwrap_or(0);
-
-    // NOTE: We need to update the stack num tokens, because the inference response might have produced
-    // less tokens than estimated what we initially estimated, from the middleware.
-    state
-        .state_manager_sender
-        .send(AtomaAtomaStateManagerEvent::UpdateStackNumTokens {
-            stack_small_id,
-            estimated_total_tokens,
-            total_tokens,
-        })
-        .map_err(|e| {
-            error!("Error updating stack num tokens: {}", e);
-            StatusCode::INTERNAL_SERVER_ERROR
-        })?;
-
-    let mut blake2b = blake2::Blake2b::new();
-    blake2b.update([payload_hash, response_hash].concat());
-    let total_hash: GenericArray<u8, U32> = blake2b.finalize();
-    let total_hash_bytes: [u8; 32] = total_hash
-        .as_slice()
-        .try_into()
-        .expect("Invalid BLAKE2b hash length");
-    state
-        .state_manager_sender
-        .send(AtomaAtomaStateManagerEvent::UpdateStackTotalHash {
-            stack_small_id,
-            total_hash: total_hash_bytes,
-        })
-        .map_err(|e| {
-            error!("Error updating stack total hash: {}", e);
-            StatusCode::INTERNAL_SERVER_ERROR
-        })?;
-    Ok(Json(response_body))
 }
 
 pub(crate) mod utils {
