+++ cursor.toml
@@ -1,0 +1,2 @@
+txDigest = "dUy6s2D8p224rypaeKyyEQ6cohmGDp9N5SAMksJfDZg"
+eventSeq = "0"
